Nearly fix bugs with special hull colors
Everything in issue #2 is resolved, unfortunately we now have the
opposite problem! The HUD is correct, the leg and scout shapes are
correct, but the head retains the true team color.

I suspect the issues lies with CBSPPart::ReplaceColor itself, based on
how it seems to usually be called. The current implementation replaces
`origColor` with `newColor`, but only if `origColor` is actually
encountered. If the color (`kMarkerColor`, for example) was already
replaced, it won't work.