fix parsing of PICT elements with DHText opcode
When porting gzr levels I encountered a bunch of lines that would split unexpectedly in the middle of variables.  For example "out[1] = @blueClose" would become "out[1] = @\rblue\rClose" which doesn't work.  Checking for element DHText and suppressing insertion of the \r character seems to fix this particular problem.  The theory is that DHText means that the text is just moved over on the same line by some horizontal amount (versus DVText which tends to be on a new line).

Note: with this change I also removed the unused (numer, denom) arguments and replaced them with the prependNewline argument.  As this is an interface change, it needs to be reviewed/approved by whoever cares about such things.