fixes related to FDistanceEstimate
The more accurate distance calculation of FDistanceEstimate messed with the Hornet's Nest & Swarm levels which depended on the old calculation.  Specifically, these levels have a recentering Teleporter which prevents Hectors from getting outside of the hexagaon shaped play area.  The deadRange of that Teleporter was too small with the more accurate distance calculation so it was easy to get into a teleportation loop.  I just increased the deadRange on these 2 levels (from 32 to 34) and all is well now.