levels and chat tab (#63)
* Allow Frameworks to also reside in ~/Library/Frameworks

* a quick little script for converting original levelsets files with resource forks to .r files

* adding 2 gzr levelsets for testing

* Fixes to build macapp target with ~/Library/Frameworks

* use more conventional dash in level set names

* removed broken levels

* Allow Frameworks to also reside in ~/Library/Frameworks

Made the parser a little more robust by ignoring lines with garbage characters.

* Fixed check for SDL2 frameworks

Oops, I introduced a bug in my previous commit on checking for where the SDL2 framework lives.  This should work now

* levels tweak so they parse better and updated Head's email address

* parsing fixes and update Head's email

* added gzr - Balledness\! levelset

* Update Makefile

Signing ID set to NONE

* fixed Glasshouse level so that it parses correctly

* unedited version of Over the Hill

* fixed parsing of Kuwait level

* more minor fixes to Kuwait level

* fixed parsing of SandStorm!

* fixed parsing of "King of the Ring" level

* fixed parsing of "Mirage" level

* fixed parsing of "Koth Showdown" level

* Revert "fixed parsing of "Koth Showdown" level"

This reverts commit 9264db2fec9cf97ca020bcc7f25a1eb4a083368c.

* Fixed parsing of "Koth Showdown" level

* multiple levels fixed

Several levels fixed and added to list in the GUI

* Print level loaded

* Print user joined server

* handle more opcodes

Some of the old gzr levels have PICTS that contains some weird opcodes.  This fix handles those better so that they are properly skipped.

* fix parsing of PICT elements with DHText opcode

When porting gzr levels I encountered a bunch of lines that would split unexpectedly in the middle of variables.  For example "out[1] = @blueClose" would become "out[1] = @\rblue\rClose" which doesn't work.  Checking for element DHText and suppressing insertion of the \r character seems to fix this particular problem.  The theory is that DHText means that the text is just moved over on the same line by some horizontal amount (versus DVText which tends to be on a new line).

Note: with this change I also removed the unused (numer, denom) arguments and replaced them with the prependNewline argument.  As this is an interface change, it needs to be reviewed/approved by whoever cares about such things.

* Save last used level

* Update CONTRIBUTORS.md

* Save last used level

* fixes related to FDistanceEstimate

The more accurate distance calculation of FDistanceEstimate messed with the Hornet's Nest & Swarm levels which depended on the old calculation.  Specifically, these levels have a recentering Teleporter which prevents Hectors from getting outside of the hexagaon shaped play area.  The deadRange of that Teleporter was too small with the more accurate distance calculation so it was easy to get into a teleportation loop.  I just increased the deadRange on these 2 levels (from 32 to 34) and all is well now.

* fix win teleporter

I think this will fix wins on Hornet, pulled it inside of the deadRange Teleporter bubble.

* remove print out

* clean up

* Parse DirectBitsRgn opcode

This fix parses the DirectBitsRgn opcode which contains a PixMap and PixData.  The PixMap parsing code could be useful for the DirectBitsRect opcode but I haven't encountered that yet.

* Added "gzr - Grecian Formula" levelset

* show server count

* import gzr R.I.P. levelset

Import gzr RIP levelset and make some small tweaks to the pict2svg.py script to make it easier to extract & convert the custom BSPs from the levelset.

* Look for bsps in levelset directory

BSPs are first searched for in the levelset's bsps directory before searching the top-level bsps directory.

* Random numbers were too big

Random numbers were 65536 times too big because they were being generated as 32-bit numbers instead of 16-bits as per the original code.

* Add RIP to the list of levelsets

* chat tab

Co-authored-by: Tom Anderson <tom@squeat.com>