added parsing of 0x90(BitsRect)
This was a journey.  I finally deduced, by looking at how ImageMagick parses the PICT file, that the 4 bytes at the beginning of the PixMap struct are skipped for opcodes that have the word "clipped" in the description.  This is not documented in the QuickDraw PDF document but the code seems to work for the few BitsRect opcodes I found.

This was primarily tested against the "Target" levels in aa-deux-normal.  Neither of these levels has a ColorTable so it would be good to find one that does so we can test that too.