Allow DefaultUDPPort to be overridden on command line
This will be helpful for quickly starting both a server and a client running on different ports.

I have configured my router to forward both ports 19567 and 19568 to forward to 19567 so that the client running on 19568 will be able to talk to the server on 19567.   I'm hoping this will be helpful in debugging crashes and hangs... we'll see.