Allow client connection from any port
Instead of pegging client connection to port 19567, allow them to be dynamically allocated by the system when creating the connection.  This makes it much easier to connect several clients on a single computer which, in turn, makes it easier to debug connection issues.

As part of this change, the -p option only sets the client port (server port settting stays in the prefs.json).