Specify port via command-line option (#83)
* Allow DefaultUDPPort to be overridden on command line

This will be helpful for quickly starting both a server and a client running on different ports.

I have configured my router to forward both ports 19567 and 19568 to forward to 19567 so that the client running on 19568 will be able to talk to the server on 19567.   I'm hoping this will be helpful in debugging crashes and hangs... we'll see.

* accept playername as command-line option -n

It's also handy to change playername on the command line when starting multiple instances.

* Allow client connection from any port

Instead of pegging client connection to port 19567, allow them to be dynamically allocated by the system when creating the connection.  This makes it much easier to connect several clients on a single computer which, in turn, makes it easier to debug connection issues.

As part of this change, the -p option only sets the client port (server port settting stays in the prefs.json).