experiment - route all packets through server
This is an ongoing change/experiment to see what happens if all packets flow through the server.  Setting the #define of ROUTE_THRU_SERVER to 1 will enable the code that sends all packets through the server.  The server then "fast forwards" any packets that are not meant for the server itself.