Server as router works!
The short version of this change is that each client sends all
requests through the server setting the senderId and distribution
in the packets so that the server can figure out what to do with the messages.
If a message is intended for another client then the server
re-routes the message to all clients specified in the distribution
field.

The tricky part of all of this is having each client keep track
of which connection it WOULD have used if it was sending the
messages directly.  So it still uses the same connection objects
as before but switchs the host/port right before sending.  Similarly,
an incoming message is immediately assigned to the correct connection
because that's where all the bookkeeping is done for that client-to-client
communication.  Messages that are not meant for the server are immediately
"fast-forwarded" to the client that it is destined for.

This still needs more testing but I was able to run games locally with
multiple clients connecting through the server.