Send ACK messages
I'm not sure how/if this ever worked because after checking if (thePacket == kPleaseSendAcknowledge) it would immediately call GetOutPacket and change thePacket.  Subsequent checks to see if it was an ACK would not have succeeded.

In my testing of the "server-as-router" feature/experiemnt, I witnessed some problems with the fact that ACKs weren't being sent.  For instance, some clients would continually resend some messages until they got an ACK back.  This fix would satisfy those requests.

It's possible that this fix might solve some one-off issues we have seen but I can't guarantee it.  But that's why I pulled it into it's own commit so that we can pull it into master and do further testing on the master branch.