focus, recents, server window, pw, network fixes (#84)
* Put 0x9b in numerical order

Just moved a bit of code up so that it's correctly ordered

* Update PICTParser.cpp

* Allow DefaultUDPPort to be overridden on command line

This will be helpful for quickly starting both a server and a client running on different ports.

I have configured my router to forward both ports 19567 and 19568 to forward to 19567 so that the client running on 19568 will be able to talk to the server on 19567.   I'm hoping this will be helpful in debugging crashes and hangs... we'll see.

* accept playername as command-line option -n

It's also handy to change playername on the command line when starting multiple instances.

* recent levels and focus

* recent levels and focus

* focus

* Allow client connection from any port

Instead of pegging client connection to port 19567, allow them to be dynamically allocated by the system when creating the connection.  This makes it much easier to connect several clients on a single computer which, in turn, makes it easier to debug connection issues.

As part of this change, the -p option only sets the client port (server port settting stays in the prefs.json).

* avoid premature packet resends

In the course of studying packet flow, I discovered that many packets were being resent before they needed to be sent.  This is a quick fix to check that the packet's nextSendTime is honored when looking for packets to send.

* focus, server window, pw, network fixes

* Send ACK messages

I'm not sure how/if this ever worked because after checking if (thePacket == kPleaseSendAcknowledge) it would immediately call GetOutPacket and change thePacket.  Subsequent checks to see if it was an ACK would not have succeeded.

In my testing of the "server-as-router" feature/experiemnt, I witnessed some problems with the fact that ACKs weren't being sent.  For instance, some clients would continually resend some messages until they got an ACK back.  This fix would satisfy those requests.

It's possible that this fix might solve some one-off issues we have seen but I can't guarantee it.  But that's why I pulled it into it's own commit so that we can pull it into master and do further testing on the master branch.

Co-authored-by: Tom Anderson <tom@squeat.com>