Change version number
It's a good idea to change the version number when the packet structure changes.  There has been a minor change to the ACK packet with the ROUTE_THRU_SERVER changes that adds senderId and distribution so that the server knows to whom the packet should be forwarded and the receiving client knows who sent it.