fix connection crashes
I noticed that it crashed often if you entered/exited a server over and over.  It turned out to be a problem with how connection were added/deleted from the socketSet.  Since each instance only needs 1 UDP connection to the world, I simplified it so that the set only ever has a single socket in the socketSet.  And it is added/delete at creation/deletion.

This seems to fix the crash I was seeing and should also get rid of the "socketset is full" error we saw in testing.