preparing server-as-router for merge to master
I decided to ditch the ROUTER_CAPABLE flag to avoid the confusion with the ROUTE_THRU_SERVER stuff.  It's always on and necessary so the ROUTER_CAPABLE flag is gone.

Also, turned OFF the ROUTE_THRU_SERVER flag by default since this is in prepartion for merging with master.

If you want to play/test with routing all messages through the server, then set the ROUTE_THRU_SERVER flag to 1 in CUDPComm.h.  This also changes the version number (from 6 to 7) so anyone connecting to your server must build with the same flag setting in order for it to work (they aren't quite compatible because of tweaks to the ACK command needed to make it work).