avoid resend avalanche
On some of our laggy game, clients can get stuck resending packets to players who are causing the lag.  It can get out of hand and the laggy player may never catch up to process  the avalanche of resent packets so that the game just spins.

This fix sets an upper limit of 2 on the number of packets that a client may re-send until it gets some sort of reply back from that laggy player.  The theory is that re-sends should be rare and are meant to account for lost packets.  If the receiving client is just slow, then it's better to hold off re-sending to that player.

I'm hoping that this will fix the "resend avalanche" problem and that laggy connections will not kill a game entirely.