Increase frame time with higher latency
This is an experiment to see what happens when we increase the frameTime as latency increases.  This will result in games slowing slightly initially but also reduces the number of packets sent.  The idea is to try to figure out if that will allow a game to recover better from a spike.