Changed AUTOLATENCYPERIOD to msec
I realized that, with a variable effective frameTime that the latency checks (via kpLatencyVote) would happen less and less frequently for higher LT values.  But it's at higher LTs that you want it to hurry up and bring the LT back down if possible.  So I redefined it to be in units of msec instead of frames so that it would continue to do the check on a consistent basis.

And just to be super-anal, I made sure that the AUTOLATENCYPERIOD variable was evenly divisible by ALL frame times (using maths).  I'm not sure that this is necessary but I wanted to avoid a situation where you switch to a lower/higher LT then it immediately does another LT check in the next frame.  It should now happen on exactly every 3840 msec interval, regardless of the current LT and frameTime.