Try to pull in LT=4 as a playable latency
While fooling around with latency I found that a latency of 4 is laggy but still a reasonable level in which a game can be played.  So let's try changing LATENCY_DIVISOR to keep the frameTime at 64ms for LT=4.  It means that recoverability might take a small hit but that people with slightly laggy connections should have a better chance of staying in the game.