Fix issue with LT changing too quickly sometimes
When switching between LTs with different frame rates, you could sometimes see the LT change less than a second later instead of the expected 3.456 seconds.  This was because the number of frames per latency vote could go from say 48 to 60 and it might check on frame 48 then again on frame 60.  I made it so that it rounds up and will skip forward to frame 120 (in that example).

I also tweaked the latency vote period to 3840 partly to slow it down a smidge but mostly to make it divisible by all the possible frame periods.  For LT < 5 you should see 60 frames (3840/64) between votes.