Improve resilience with packet loss
This change affects when urgent (game time) packets are resent.  Previously they weren't sent in a very urgent manner which is a problem with connections that lose packets.  This change attempt to resend packets so the resent packets have a chance of getting there in the correct frame.  This makes the game smoother at the cost of a higher resend rate... probably more like what the original game did.