Use overall Latency to determine urgent resend
Using an estimate of overall latency allows faster clients to reduce resending of packets to each other and (hopefully) only the clients causing the latency will have to do most of the resends.  But it can also reduces the urgentRestransmitTime for a low latency game making those games feel faster if a client has packet loss (maybe?).