Add `invertYAxis` setting, start `joystickMode`
The setting for `joystickMode` will apply correctly if the relevant code
blocks are uncommented, however, I'm unsure what `joystickMode` is
actually supposed to do and looking at Juri's original source did not
illuminate this much for me. All I know is that it flips the `doCenter`
parameter on `DoMouseControl`. Thus, I've left this for somebody else to
finish up when the need arises. It should probably be a separate issue,
anyway.

Fixes #160