Move/rename Choke Daddy B, cleanups/fixes
All Air Voss variants have been massaged into consistency and now live
in the Quirks Mode collection, due to the altered weapons. I've also
changed the masking on a ton of "destination-only" teleporters so that
players cannot accidentally be caught in their gravity wells. (This was
mostly only an issue on the low gravity Air Voss variants, but it's a
good measure to apply everywhere just in case.)