Make sure GameTick() runs next frame in tests
The previous chipmunk fix exposed a slight issue with testing.
The reason GameTick() would always advance to the next frame in
tests was similar to why we would experience chipmunk mode in games.
This fix makes sure that every call to TestGame::GameTick() will
force the next frame to be processed.