fixed jumping bug
Sometimes you could get a kfuJump UP and a kfuJump DOWN in the same frame resulting in a big jump (maximum crouch).  But, if you only detected the kfuJump UP in a single frame, it would cut the crouch in half before calculating the jump velocity.   The fix is simple, execute the UP handler first before the crouch can be cut in half.