fix stick-the-landing bug... maybe
I gotta be honest, I don't like this fix.  It seems to work but I don't fully understand the MoveLegs code.  Anyways, the Hector was bouncing too far into the ground and this fix seems to help limit the downward speed after it gets into bounce mode.  But I'm a little uncomfortable with the fix so we'll have to test the shit out of it!