Add more robust testing for HECTOR energy levels
We had isolated tests for checking shield, energy, and plasma energy
levels, but were not acknowledging the relationship between the three.
Every one of the previous shield/energy/plasma tests are now checking
the expected values for shields AND energy AND plasma over time, an
additional test was added to check plasma recharge with a booster
active, and yet another more complex test was added to verify
regeneration rates when NONE of these energy levels start out at full
power (which should result in slower regeneration across the board).

I also replaced my usage of std::tuple with a new class,
HectorEnergyReadings, to store a snapshot of all four energy levels in a
given frame.