prevent certain false-positive fragment errors
It is possible to receive a latency vote packet after the vote has been counted, especially during high latency or packet loss.  This could have resulted in a fragment being recorded and used on the next latency vote and display a fragment error even though there was none.  This fix adds logic to narrow the window of consideration for fragments to the current voting window.