let's pretend like this never happened
This is the biggest mistake I made on the Avara code (about a year ago).  I was confusing a packet-loss problem with a latency problem.  And this code added way more complexity than it was worth.  It has been effectively disabled for a little while and it hasn't affected game play at all... so I'm gonna just delete it and hope nobody notices.