make tests pass without a config file
There are 2 things going on here.  First, moving the call of ReadPrefs() from to the implementation (.cpp) file fixes the ability to run Avara without a config file.  I can make some sense of this... maybe it's related to timing of the loading of the config file.  But it doesn't fix the test!  However, renaming the variable 'prefs' to '_prefs' fixes the test.  WTF?!  I don't understand why changing the variable name works but it does... can anyone explain this?  Is there a 'prefs' variable somewhere else in the code that this conflicts with?