Trying to reduce packet resends
Currently the number of "resend" packets represents between 33% and 50% of all packets sent. I think this is because the urgentRetransmitTime is often much lower than it needs to be so it resends almost everything.
With this change I'm trying to do a better job of computing the last possible time to re-send a game packet and still have it arrive close to the target time (2*LT) on the slowest connection.  On my local machine I see resend rates drop below 20% and sometimes 10% with no apparent degradation in game play.  We'll see if the real-world results are as good.