Fix getting stuck on ledges and blocks
Initially I didn't think I needed to scale MINSPEED because speeds are in "classic" units.  What I failed to realize is that this is a check against small speeds and speeds can start out small for high-FPS games (like the first few frames trying to back off of an object).