Restore regular/super jump capability
Instead the super jump being an undocumented feature that depended on a player hitting the jump button on the exact frame when another jump button was release, this makes it more of a documented feature where the superjump happens when one jump key is in the held (down) position while releasing another jump key.  This also makes the super jump behavior more consistent across frame rates.