Rewrote jump fix using bitmask instead of map
I did this for a couple reasons.  First, it's not necessary to keep an accurate count of how many keys are pressed, just that there are more than 1 keys pressed for a particular function.  Second, if we ever want/need to send this as part of the message it matches well with how we currently send up/down/held data now so it would be an easy addition.