Adding functionality to record & playback of the current player
The two new command-line options allow the current game-play to be record to stdout if the -o or --keys-to-stdout option is specified.  Also, you can replay the movements recorded in a previous session with the -i or --keys-from-stdin option.  For example to playback AND record you could run Avara like this:

> cat recording.1 | Avara -i -o > recording.2