messing with Baghdad 3, moved to quirks-mode
I might have gone too far BUT...
I wanted to try setting the goody size relative to the weapon power as a visual cue that show the relative strength.
In this version, if grenades are more powerful, then missiles are weaker and vice versa.  Also, when a weapon is more powerful, there are fewer of them in the Goody and vice versa.
And of course, it's all random so you can get big missiles or big grenades (but not both).