eval all the attributes (#238)
* mask="0" on recenter teleporter and cleanup

* Allow almost any attribute to be an expression

This change to the parser was motivated by the desire to have more flexibility in
specifying object & goody locations.  The Baghdad 3 level was modified to take advantage
of these changes with dynamically placed goodies and random visibility of the floor
boundary (for Seven).

* removed debugging stuff

* Fix double/float conversion to Fixed

Conversions with doubles and floats can give different Fixed values.  It's probably best to use ToFixed everywhere.

* fix RandomIncarnator logic

In previous fix (7b95802) to handle no incarnators on a level, I broke
the case of not enough incarnators.  With this fix, I completely rewrote
CAbstractPlayer::Reincarnate to make it easier to understand so hopefully
it will be less likely to be broken again.

* use ToFloat(x) instead of x/65536.0

* messing with Baghdad 3, moved to quirks-mode

I might have gone too far BUT...
I wanted to try setting the goody size relative to the weapon power as a visual cue that show the relative strength.
In this version, if grenades are more powerful, then missiles are weaker and vice versa.  Also, when a weapon is more powerful, there are fewer of them in the Goody and vice versa.
And of course, it's all random so you can get big missiles or big grenades (but not both).

* level tweaks

* move goody to eye level
* increased max scale to 3x

* improved & quantized goodie scaling

It should be normal 1x scaling about 50% of the time, 2x = 36% and 3x = 14%.  Quantized the scaling to make it more predictable.

* Show more of the smaller scale goodies

Mostly just playing with visualization but when the grenades are 2x size, then you will see 2 half-size grenade goodies at each spot and vice versa.
Gives players a quick indication of whether the level is 1x, 2x, or 3x version.

1x version appears 50% of the time
2x versions appear 36%
3x versions appear 14%

* revert to loading random level at startup

* tweak to set minimim random incarnator radius