fix auto-pause by not doing it
Instead of pausing automatically, which is a big change, I fixed the ability to abort from a hung game by pressing the kfuAbortGame key.  So it's up to the user to abort out of a hung game.

Spectators will automatically abort out of a game in which they receive so many frames that it wraps around the functions buffer.