transmit sendCount as part of the packet
This will help us to better understand how often the first packet is not received by a client.
Also modified the Roster to display a smoothed average of the lost packets as estimated by how often we don't receive the packet with sendCount==0.  So the roster now shows {$Name ($RTT/$Loss%)}.