FrameNumber with a sprinkling of FrameTime
I tried to find any variable that was used in frameNumber calculations and converted them all to type FrameNumber which is `typedef int32_t`.  Now everything dealing with frame numbers should have the same type and 1 place to change that type.

Also did frameTime but there are not very many of those variables.

Killed a few warnings along the way.