Merge the record/replay prototype (#243)
* Adding functionality to record & playback of the current player

The two new command-line options allow the current game-play to be record to stdout if the -o or --keys-to-stdout option is specified.  Also, you can replay the movements recorded in a previous session with the -i or --keys-from-stdin option.  For example to playback AND record you could run Avara like this:

> cat recording.1 | Avara -i -o > recording.2

this is to allow external clients join a game with clients running on the same host as the server (e.g. a server with lots of bots running locally)

* loop.sh script to loop over recorded keystrokes

* rewrite connect host to localhost if the server thinks we're the same

* just show connection table as received and after rewrites

* ClockType is an integer timer returned by GetClock()