get spectator mode (/spectate) working!
I kinda took the "easy" solution and just killed off the spectator players on the first frame.  This can result in an explosion of the spectator hectors (or we can comment that out if people don't like it).
Also gave the loading status enum the name `LoadingState` so that the state name shows up in the debugger if you type the variable with that.