fixes spectator losing touch after abort
The problem was that, if a packet was lost and the sender had moved on and couldn't re-send it, the packet queue would continue to grow and possibly never be cleaned up.  This change allows the non-players to be allowed to skip over a lost packet in such a way that the queues would then be processed and cleaned up.

Also, I finally figured out how to reliably cause this bug by deleting not only the frame packet, but also to skip the serialNumber of that packet forward as if it had been lost.

I was playing with Ping thinking that might help to send a packet to trigger the cleanup.  It turned out that didn't help this particular problem, but I like the refactor that makes inactive players fewer packets to active players.