allow all players to skip forward on missing packets
It turns out to be better if you just let all players, active or not, to skip forward when a packet is missing.  There's a good chance that packet has already been resent and in the worst case it will be asked for again in the wait loop.  Now that wait loop exits earlier because it is effectively looking further down the queue for the right frame number.