fix flawed wait loop
the GetFunction() wait loop, the one that can abort you, had a fundamental flaw,
it would send a message to re-send a particular frame but that message would
end up on the end of a read queue that was still waiting for the original message
that caused the wait loop in the first place.  This change rewrites some of that
logic to clean-up the queue when it enters this wait loop.  This should mean that
we'll see fewer actual aborts with an occasional hiccup while it is request and
getting the needed frame.