ask for some re-send packets a little faster
When the next frame isn't here yet, you wait a bit but the previous wait seems a bit too much for a 16ms frame.  I reduced it a bit and this seems a little better.  It might make the game feel a little smoother when the spikes hit?  Hard to know for sure.