Allow HUD color to be customized
The old `weaponSightAlpha` pref has also been renamed to `hudAlpha` to
better match this new `hudColor` setting.

As part of these changes, ColorManager now gets updated whenever there
is a prefs change. This also means that I was able to update the team
color dropdowns whenever color blind mode is changed, without the player
having to restart the application.

Changes to `hudColor`, `hudAlpha`, and `colorBlindMode` are not applied
to games-in-progress, however, and it is probably not worth attempting
to change that--at least for color blind mode. The HUD could be updated
but we would need to destroy and recreate the dashboard and weapon sight
elements.