more accurate RTT calculation
The last ping command sent can't be relied on for the RTT calculation because the receiver may not respond to it right away.  This is especially important for the RTT calculation displayed outside of the game action.

Since the RTT calculation is more accurate, sending fewer pings overall, more frequently and weighting them more in the calculation.