scoreKeeper remembers players who leave game
I noticed that when a player leaves a server after dying, the ScoreKeeper wouldn't know who was playing or it might mis-attribute the game to a newly joined player.  This fix makes the ScoreKeeper remember who started the game in each position.
This also allows the Elo calculation to get the finishers correct for player ratings.