reduce auto-LT churn
Biggest change is ceil on RoundTripToFrameLatency() calculation instead of round.  That helps to prevent LT being set too low then having to increase right away.

Also increased period for how often LT can be reduced.  And tweaked threshold for when to add 1 to LT so that it doesn't immediately go back up to previous LT as often.