Standardize KOTH and Captures units & messaging
Instead of reporting time in seconds in KOTH, I've moved to using
"points" as I was in captures. I've also standardized reporting
intervals to every ten points (was previously fifteen in Captures), with
a max goal of 100. However, the true value of these "points" is still
scaleable under the hood via the `pReqPointMultiplier` parameter, which
was the main motivation for this change on the KOTH side. Both game
modes also underwent minor simplifications to reduce the number of game
objects in use.

Unscaled KOTH games still require a team to be on the hill for 100
seconds in order to win, so functionally there is no gameplay
difference.

Unscaled Captures games now acquire 10 points for every ~40 seconds a
single capture point is being held, so more or less exactly 4x slower
than unscaled KOTH games. This seems appropriate given you don't need to
be standing on it to accrue time, and there are likely multiple capture
points per level.