Game builds, level set list replaced w/ new impl
I've somewhat separated the concept of an asset "repository" (which
provides a listing of available level sets) from asset "storage" (where
files are stored locally). The idea is that a "RemoveAssetRepository"
type may hold a reference to local storage, where downloaded assets will
actually be saved.

Lastly...the game builds!