Fix avarascript the least efficient way possible
As I suspected the C-string pointer was being invalidated, occasionally
leading to garbage data. This stupid version just copies the string's
inner C-string data into a separate character array that lasts as long
as it needs to in order to run the script. There has to be a better way,
but even loading the script into a C-string initially seemed to present
the same bugs as I was seeing before.