avoid frags caused by dead player aborting
When a player aborted a game after their last life their client would send an asynchronous message letting other players knowing they are out.  But this could sometimes result in a race condition where other players wouldn't be able to handle the abort keystroke which allows it to clean up and set the `isOut` flag.  The result of this was that the aborted player would be counted by some players and not others which messed up the fragmentation checksum.