more frag fixes
The crux of abort/frag problems is/was that the notification of the abort was asynchronous.  But the other players in the game don't need to get the async message because they will get the abort key press so stop sending that message to active players.

Also, there's no good reason to keep updating the frag checksum for players who are out of the game.  This was also related to the previous problem where players were adding in the location checksum and stopping that addition at different times.  This fix is really extra insurance that an aborted or dead player won't cause frags.