try taking the hill logo out of the equation
the solid logo in the middle of the hill seems to be the only difference
between this level and the other levels that don't frag.  Temporarily elevating
them out of the way.  Also I used Holograms for some but the Hologram doesn't
respect "pitch".  If Holograms work, we might consider adding pitch to CHologramActor
similar to CSolidActor.