make FramesFromNow() accept a float arg
Well, this is embarrassing.

Most of the time FrameFromNow() is passed an integer.  But there's an important call that passes a float on this check:
    while (FramesFromNow(latencyTolerance) > topSentFrame) {
        itsNet->FrameAction();
    }

Passing latencyTolerance as an integer was truncating the LT.  For example, an LT=1.5 would be treated as LT=1 and result in
less than the full LT of frames being sent on time.