fix level-loading crashes
When new levels are loaded the Hull and Level/Set menus are re-created from scratch.  This would leave dangling pointers in nanogui::Screen (mDragWidget and mFocusPath).  The "best" fix would probably be to replace all Widget pointers with shared_ptr<Widget>.  But this is a quick fix that gets the job done.

I also added "/dbg rload X" which will load a new level every X seconds.  This was useful for recreating this bug but also might be useful for discovering other random level-load issues in the future so I left it in.